Added a minimaxAI
Created a minimaxAI. First created a isMovesLeft to check if there are
any moves left. Then created evaluate function to evaluate the condition
of the board. Later added a minimax function to check the moves ahead.
Then added a findTheBestMove function to use the minimax function and
find the best move available. Later added makeBestMove function to make
best move.